Removed Campaign Added Age Renamed crisis die to ds, not dc (to allow coin flips) Continued work on NPC sheet.
