+++ template.json
@@ -10,7 +10,6 @@
                     "concept": "",
                     "experience": 0,
                     "species": "",
-                    "campaign": "",
                     "ambition": "",
                     "faction": "",
                     "wealth": 0,
