+++ templates/partials/sheet-tabs/actor-info.hbs
@@ -18,11 +18,11 @@
                         <td>
                             <input name="name" type="text" value="{{actor.name}}" placeholder="{{localize "c2d10.character.name"}}" />
                         </td>
-                        <th>
-                            {{localize "c2d10.character.campaign"}}
+                        <th class="cd10-header">
+                            {{localize "c2d10.sheet.age"}}
                         </th>
                         <td>
-                            <input name="system.info.campaign" type="text" value="{{system.info.campaign}}" placeholder="{{localize "c2d10.character.campaign"}}" />
+                            <input name="system.info.age" type="text" value="{{system.info.age}}" placeholder="{{localize "c2d10.sheet.age"}}" />
                         </td>
                     </tr>
                     <tr>
