+++ templates/sheets/npc-sheet.hbs
@@ -8,41 +8,21 @@
                 <h3>
                     {{localize "c2d10.sheet.skills"}}
                 </h3>
-                <div class="content-box c2d10-subsection flex-row flex-between">
-                    <div class="flex-col flex-start w-100">
-                        <div class="c2d10-contentbox {{#if showEffects}}fancy-effects{{/if}}">
-                            <table>
-                                <colgroup>
-                                    <col class="w-60">
-                                    <col class="w-40">
-                                </colgroup>
-                                <tr>
-                                    <th class="c2d10-test c2d10-skill-test table-no-border description"
-                                        data-tooltip="{{localize (concat "c2d10.descriptions." system.skills.primary.name)}}
-                                        <p>Left click to roll, right click or shift+click for description.</p>"
-                                        data-type="skills" data-group="npc"
-                                        data-id="{{system.skills.primary.name}}"
-                                    >
-                                        {{localize (concat "c2d10.skills." system.skills.primary.name)}}
-                                    </th>
-                                    <td class="flex-row table-no-border resource-row" data-type="skills" data-group="npc" data-id="{{name}}" data-tooltip="{{localize "c2d10.sheet.dot-tooltip"}}">
-                                        {{#dots system.skills.primary.level 5}}{{/dots}}
-                                    </td>
-                                </tr>
-                                <tr>
-                                    <th class="c2d10-test c2d10-skill-test table-no-border description"
-                                        data-tooltip="{{localize (concat "c2d10.descriptions." system.skills.secondary.name)}}
-                                        <p>Left click to roll, right click or shift+click for description.</p>"
-                                        data-type="skills" data-group="npc"
-                                        data-id="{{system.skills.secondary.name}}"
-                                    >
-                                        {{localize (concat "c2d10.skills." system.skills.secondary.name)}}
-                                    </th>
-                                    <td class="flex-row table-no-border resource-row" data-type="skills" data-group="npc" data-id="{{name}}" data-tooltip="{{localize "c2d10.sheet.dot-tooltip"}}">
-                                        {{#dots system.skills.secondary.level 5}}{{/dots}}
-                                    </td>
-                                </tr>
-                            </table>
+                <div class="content-box c2d10-subsection {{#if showEffects}}fancy-effects{{/if}} flex-row flex-between">
+                    <div class="c2d10-contentbox crisis-box w-50 flex-row flex-between">
+                        <div class="c2d10-test c2d10-skill-test description" data-type="skills" data-id="{{system.skills.primary.name}}" data-level="{{system.skills.primary.level}}" data-tooltip="<p>Left click to roll.</p>">
+                            <input name="system.skills.primary.name" type="text" value="{{system.skills.primary.name}}" placeholder="{{localize "c2d10.sheet.skill"}}"/>
+                        </div>
+                        <div class="flex-row flex-center resource-row" data-type="skills" data-id="primary" data-pass="false">
+                            {{#dots system.skills.primary.level 5}}{{/dots}}
+                        </div>
+                    </div>
+                    <div class="c2d10-contentbox crisis-box w-50 flex-row flex-between">
+                        <div class="c2d10-test c2d10-skill-test description" data-type="skills" data-id="{{system.skills.primary.name}}" data-level="{{system.skills.secondary.level}}" data-tooltip="<p>Left click to roll.</p>">
+                            <input name="system.skills.secondary.name" type="text" value="{{system.skills.secondary.name}}" placeholder="{{localize "c2d10.sheet.skill"}}"/>
+                        </div>
+                        <div class="flex-row flex-center resource-row" data-type="skills" data-id="secondary" data-pass="false">
+                            {{#dots system.skills.secondary.level 5}}{{/dots}}
                         </div>
                     </div>
                 </div>
